added modify function

+++ src/main/java/org/example/bokningssystem/repo/BokningRepo.java
@@ -3,5 +3,5 @@ package org.example.bokningssystem.repo;
 import org.example.bokningssystem.modell.Bokning;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BokningRepo extends JpaRepository<Bokning, Integer> {
+public interface BokningRepo extends JpaRepository<Bokning, Long> {
 }
